refactor(slack-bridge): extract Pinet helper seam
