+++ pinet-core/README.md
@@ -1,5 +1,13 @@
 # @gugu910/pi-pinet-core
 
-Scaffold package for the planned extraction of broker/runtime/Pinet core code out of `slack-bridge`.
+Runtime-core helpers for Pinet that are independent of any Slack adapter.
+
+Current seam:
+
+- Pinet output option normalization (`cli` default, explicit `json`/`full` opt-ins)
+- durable Pinet read result text/detail formatting
+- scheduled wake-up time parsing and thread ID helpers
+
+`@gugu910/pi-slack-bridge` still composes the extension and preserves compatibility wrappers, but these helpers now live behind package exports so future extraction can move one boundary at a time.
 
 Design proposal: `plans/slack-split-proposal.md`
