+++ pinet-core/index.ts
@@ -1,4 +1,3 @@
-// Scaffold package for the planned Pinet runtime-core extraction.
-// See plans/slack-split-proposal.md for the proposed module boundary.
-
-export {};
+export * from "./output-options.js";
+export * from "./pinet-read-formatting.js";
+export * from "./scheduled-wakeups.js";
