+++ slack-bridge/pinet-tools.ts
@@ -1,8 +1,16 @@
 import os from "node:os";
 import {
-  classifyPinetMail,
-  formatPinetMailClassLabel,
-} from "@gugu910/pi-broker-core/mail-classification";
+  buildCompactPinetReadDetails,
+  formatPinetReadResultCompact,
+  formatPinetReadResultFull,
+  type PinetReadOptions,
+  type PinetReadResult,
+} from "@gugu910/pi-pinet-core/pinet-read-formatting";
+import {
+  normalizePinetOutputOptions,
+  type PinetOutputOptions,
+} from "@gugu910/pi-pinet-core/output-options";
+import { resolveScheduledWakeupFireAt } from "@gugu910/pi-pinet-core/scheduled-wakeups";
 import type { ExtensionAPI, ExtensionContext } from "@mariozechner/pi-coding-agent";
 import { Type } from "@sinclair/typebox";
 import {
@@ -15,8 +23,6 @@ import {
 import { isBroadcastChannelTarget } from "./broker/agent-messaging.js";
 import { DEFAULT_HEARTBEAT_TIMEOUT_MS } from "./broker/socket-server.js";
 import { HEARTBEAT_INTERVAL_MS } from "./broker/client.js";
-import type { PinetReadOptions, PinetReadResult } from "./broker/client.js";
-import { resolveScheduledWakeupFireAt } from "./scheduled-wakeups.js";
 
 export interface PinetToolsAgentRecord {
   emoji: string;
@@ -78,13 +84,6 @@ type PinetDispatcherErrorClass = "input" | "state" | "runtime" | "network";
 
 type PinetDispatcherStatus = "succeeded" | "failed";
 
-type PinetOutputFormat = "cli" | "json";
-
-interface PinetOutputOptions {
-  format: PinetOutputFormat;
-  full: boolean;
-}
-
 interface PinetToolResult {
   content: Array<{ type: "text"; text: string }>;
   details?: unknown;
@@ -139,27 +138,6 @@ const PINET_OUTPUT_OPTION_PARAMETERS = {
     }),
   ),
 };
-
-function normalizePinetOutputOptions(args: Record<string, unknown>): PinetOutputOptions {
-  const rawFormat = args.format ?? args.f ?? args["-f"];
-  const format = rawFormat == null ? "cli" : String(rawFormat).trim().toLowerCase();
-  if (format !== "cli" && format !== "json") {
-    throw new Error('format must be "cli" or "json".');
-  }
-
-  const rawFull = args.full ?? args["--full"];
-  if (rawFull != null && typeof rawFull !== "boolean") {
-    throw new Error("full must be a boolean when provided.");
-  }
-
-  return { format, full: rawFull === true };
-}
-
-function truncateText(value: string, maxLength = 180): string {
-  const collapsed = value.replace(/\s+/g, " ").trim();
-  if (collapsed.length <= maxLength) return collapsed;
-  return `${collapsed.slice(0, Math.max(0, maxLength - 1))}…`;
-}
 
 function getRecordString(
   record: Record<string, unknown> | null | undefined,
@@ -349,116 +327,6 @@ function buildPinetAgentsHintText(hint: PinetAgentsRoutingHint): string {
   ]
     .filter((item): item is string => Boolean(item))
     .join(" · ")}`;
-}
-
-function formatPinetReadResultFull(result: PinetReadResult, options: PinetReadOptions): string {
-  const scope = options.threadId ? `thread ${options.threadId}` : "your Pinet inbox";
-  const mode = options.unreadOnly === false ? "latest" : "unread";
-  const lines = [
-    `Pinet read (${mode}) from ${scope}: ${result.messages.length} message${result.messages.length === 1 ? "" : "s"}.`,
-    `Unread before: ${result.unreadCountBefore}; unread after: ${result.unreadCountAfter}.`,
-  ];
-
-  if (result.messages.length > 0) {
-    lines.push("");
-    for (const item of result.messages) {
-      const classification = classifyPinetMail({
-        source: item.message.source,
-        threadId: item.message.threadId,
-        sender: item.message.sender,
-        body: item.message.body,
-        metadata: item.message.metadata,
-      });
-      const label = formatPinetMailClassLabel(classification.class);
-      lines.push(
-        `- [${label}] [${item.message.source}/${item.message.threadId} #${item.message.id}] ${item.message.sender}: ${item.message.body}`,
-      );
-    }
-  }
-
-  if (result.unreadThreads.length > 0) {
-    lines.push("", "Unread thread pointers:");
-    for (const thread of result.unreadThreads.slice(0, 10)) {
-      const label = formatPinetMailClassLabel(thread.highestMailClass);
-      const counts = [
-        thread.mailClassCounts.steering > 0 ? `${thread.mailClassCounts.steering} steering` : null,
-        thread.mailClassCounts.fwup > 0 ? `${thread.mailClassCounts.fwup} fwup` : null,
-        thread.mailClassCounts.maintenance_context > 0
-          ? `${thread.mailClassCounts.maintenance_context} maintenance/context`
-          : null,
-      ]
-        .filter((item): item is string => Boolean(item))
-        .join(", ");
-      lines.push(
-        `- [${label}] ${thread.threadId} (${thread.source}${thread.channel ? `/${thread.channel}` : ""}): ${thread.unreadCount} unread${counts ? ` (${counts})` : ""}; latest #${thread.latestMessageId}; pointer=pinet action=read args.thread_id=${thread.threadId} args.unread_only=true`,
-      );
-    }
-  }
-
-  if (result.markedReadIds.length > 0) {
-    lines.push("", `Marked read: ${result.markedReadIds.join(", ")}.`);
-  }
-
-  return lines.join("\n");
-}
-
-function summarizeUnreadThreadCounts(thread: PinetReadResult["unreadThreads"][number]): string {
-  return [
-    thread.mailClassCounts.steering > 0 ? `${thread.mailClassCounts.steering} steering` : null,
-    thread.mailClassCounts.fwup > 0 ? `${thread.mailClassCounts.fwup} fwup` : null,
-    thread.mailClassCounts.maintenance_context > 0
-      ? `${thread.mailClassCounts.maintenance_context} maintenance/context`
-      : null,
-  ]
-    .filter((item): item is string => Boolean(item))
-    .join(", ");
-}
-
-function buildCompactPinetReadDetails(result: PinetReadResult): Record<string, unknown> {
-  return {
-    messageCount: result.messages.length,
-    unreadCountBefore: result.unreadCountBefore,
-    unreadCountAfter: result.unreadCountAfter,
-    markedReadIds: result.markedReadIds,
-    messages: result.messages.map((item) => {
-      const classification = classifyPinetMail({
-        source: item.message.source,
-        threadId: item.message.threadId,
-        sender: item.message.sender,
-        body: item.message.body,
-        metadata: item.message.metadata,
-      });
-      return {
-        inboxId: item.inboxId,
-        messageId: item.message.id,
-        threadId: item.message.threadId,
-        source: item.message.source,
-        sender: item.message.sender,
-        class: classification.class,
-        preview: truncateText(item.message.body),
-      };
-    }),
-    unreadThreads: result.unreadThreads.slice(0, 10).map((thread) => ({
-      threadId: thread.threadId,
-      source: thread.source,
-      unreadCount: thread.unreadCount,
-      latestMessageId: thread.latestMessageId,
-      highestMailClass: thread.highestMailClass,
-      mailClassSummary: summarizeUnreadThreadCounts(thread),
-    })),
-  };
-}
-
-function formatPinetReadResultCompact(result: PinetReadResult, options: PinetReadOptions): string {
-  const mode = options.unreadOnly === false ? "latest" : "unread";
-  const markedSuffix =
-    result.markedReadIds.length > 0 ? `; marked ${result.markedReadIds.length}` : "";
-  const unreadThreadSuffix =
-    result.unreadThreads.length > 0
-      ? `; ${result.unreadThreads.length} unread thread${result.unreadThreads.length === 1 ? "" : "s"}`
-      : "";
-
-  return `Pinet read: ${result.messages.length} ${mode} message${result.messages.length === 1 ? "" : "s"}; unread ${result.unreadCountBefore}→${result.unreadCountAfter}${markedSuffix}${unreadThreadSuffix}.`;
 }
 
 function runPinetSendAction(
